fix submitted string date and add new date format,and move object to data package ,remove code example
