+++ build.gradle.kts
@@ -13,6 +13,7 @@ dependencies {
     testImplementation(kotlin("test"))
     implementation("io.insert-koin:koin-core:4.0.2")
     implementation ("com.opencsv:opencsv:5.6")
+    implementation("org.jetbrains.kotlinx:kotlinx-datetime:0.5.0")
 }
 
 tasks.test {
